+++ ingenialink/canopen/register.py
@@ -1,3 +1,5 @@
+from typing import Optional, Any, Union, Tuple, Dict, List
+
 from ingenialink.register import Register
 from ingenialink.enums.register import REG_DTYPE, REG_ACCESS, REG_PHY, REG_ADDRESS_TYPE
 
@@ -6,23 +8,23 @@ class CanopenRegister(Register):
     """CANopen Register.
 
     Args:
-        idx (int): Index of the register.
-        subidx (int): Subindex of the register.
-        dtype (REG_DTYPE): Data type.
-        access (REG_ACCESS): Access type.
-        identifier (str): Identifier.
-        units (str): Units.
-        cyclic (str): Cyclic typed register.
-        phy (REG_PHY, optional): Physical units.
-        subnode (int): Subnode.
-        storage (any, optional): Storage.
-        reg_range (tuple, optional): Range (min, max).
-        labels (dict, optional): Register labels.
-        enums (dict): Enumeration registers.
-        cat_id (str, optional): Category ID.
-        scat_id (str, optional): Sub-category ID.
-        internal_use (int, optional): Internal use.
-        address_type (REG_ADDRESS_TYPE): Address Type.
+        idx: Index of the register.
+        subidx: Subindex of the register.
+        dtype: Data type.
+        access: Access type.
+        identifier: Identifier.
+        units: Units.
+        cyclic: Cyclic typed register.
+        phy: Physical units.
+        subnode: Subnode.
+        storage: Storage.
+        reg_range: Range (min, max).
+        labels: Register labels.
+        enums: Enumeration registers.
+        cat_id: Category ID.
+        scat_id: Sub-category ID.
+        internal_use: Internal use.
+        address_type: Address Type.
 
     Raises:
         TypeError: If any of the parameters has invalid type.
@@ -33,23 +35,25 @@ class CanopenRegister(Register):
 
     def __init__(
         self,
-        idx,
-        subidx,
-        dtype,
-        access,
-        identifier=None,
-        units=None,
-        cyclic="CONFIG",
-        phy=REG_PHY.NONE,
-        subnode=1,
-        storage=None,
-        reg_range=(None, None),
-        labels=None,
-        enums=None,
-        cat_id=None,
-        scat_id=None,
-        internal_use=0,
-        address_type=None,
+        idx: int,
+        subidx: int,
+        dtype: REG_DTYPE,
+        access: REG_ACCESS,
+        identifier: Optional[str] = None,
+        units: Optional[str] = None,
+        cyclic: str = "CONFIG",
+        phy: REG_PHY = REG_PHY.NONE,
+        subnode: int = 1,
+        storage: Any = None,
+        reg_range: Union[
+            Tuple[None, None], Tuple[int, int], Tuple[float, float], Tuple[str, str]
+        ] = (None, None),
+        labels: Optional[Dict[str, str]] = None,
+        enums: Optional[List[Dict[str, Union[str, int]]]] = None,
+        cat_id: Optional[str] = None,
+        scat_id: Optional[str] = None,
+        internal_use: int = 0,
+        address_type: Optional[REG_ADDRESS_TYPE] = None,
     ):
         super().__init__(
             dtype,
@@ -73,16 +77,16 @@ class CanopenRegister(Register):
         self.__subidx = subidx
 
     @property
-    def idx(self):
-        """int: Register index."""
+    def idx(self) -> int:
+        """Register index."""
         return self.__idx
 
     @property
-    def subidx(self):
-        """int: Register subindex."""
+    def subidx(self) -> int:
+        """Register subindex."""
         return self.__subidx
 
     @property
-    def mapped_address(self):
-        """int: Register mapped address used for monitoring/disturbance."""
+    def mapped_address(self) -> int:
+        """Register mapped address used for monitoring/disturbance."""
         return self.idx